requirements.txt
flask
flask-sqlalchemy
flask-login
google-genai
python-dotenv
werkzeug
gunicorn
pillow
pytesseract
requests
